# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Portal API
  version: 0.1.0
  description: |
    Portal — self-hosted media platform API.

    This document is the **source of truth** for the HTTP contract.
    Server stubs (Go) and the TypeScript client are both generated from it.

    ## Errors

    Error responses use the **RFC 7807** `application/problem+json` shape
    (`#/components/schemas/Problem`), and that is the only error shape on the
    wire. The stable `type` URI is both the i18n key and the machine-stable
    slug: a domain code becomes `<module>/<code>` with underscores hyphenated
    (`email_taken` → `account/email-taken`), which is what the frontend's
    problem catalog keys on. Transport-level codes that add nothing to the
    status line (`unauthorized`, `forbidden`, `bad_request`, `internal`,
    `not_found`) collapse to `about:blank`.

    > **Migration status (v1):** the RFC 7807 retrofit per [D-7]/[D-29] is
    > **complete** — every handler, the auth/RBAC middleware, the rate limiter
    > and the `/continue` route now answer with `Problem`. The legacy
    > `{code, message}` body (`#/components/schemas/Error`) is emitted nowhere
    > and referenced by no operation; the schema is kept `deprecated` so a
    > generated client that still names the type keeps compiling, and should be
    > dropped at the next breaking release.

    Auth is **local password** ([ADR-06](../docs/adr/06-local-auth-model.md));
...
servers:
...
  # in any spec viewer hit a 404 on everything.
...
paths:
...
  /assets/{id}:
    get:
      tags: [assets]
      summary: Get asset metadata
      operationId: getAsset
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetID"
      responses:
        "200":
          description: Asset (includes `hlsUrl` once status is `ready`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [assets]
      summary: Update an asset (visibility)
      description: |
        Changes the mutable parts of an asset. Today that is `visibility` alone —
        the switch between owner-only and readable by anyone with the link.

        Restricted to the owner or an admin of the owning tenant, and enforced by
        the `0032` UPDATE policy rather than by a check in the handler. Someone
        else's asset is therefore reported as 404, never 403: the endpoint must
        not confirm that an id exists.
      operationId: patchAsset
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                visibility:
                  $ref: "#/components/schemas/AssetVisibility"
      responses:
        "200":
          description: The updated asset
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset:
                    $ref: "#/components/schemas/Asset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: "`media/invalid-visibility` — not one of `private` / `public`"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags: [assets]
      summary: Delete an asset (owner, or `assets:delete:any`)
      description: |
        Soft-deletes (`status='deleting'`, hidden from listings) then purges the
        asset's storage objects (original, HLS, variants) and rows. Enforced via
        `RequireOwnerOrPermission`. Idempotent — an already-gone asset returns 404.
      operationId: deleteAsset
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssetID"
      responses:
        "204": { description: Deleted }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: |
        Shared secret (COMIC_SYNC_SECRET) for service-to-service calls from the
        Python scraper. These endpoints carry no user session, which is why each
        one also takes owner_id — see the SPEC-02 P1.8 sync flow.
  parameters:
    AssetID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated, but lacks permission (or step-up required)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 problem detail, served as `application/problem+json`. `type` is a
        stable URI (also the i18n key); Portal adds the extensions below.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: "about:blank"
          description: Stable problem-type URI (i18n key).
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code, repeated for convenience.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: |
            Reserved. The machine-stable slug lives in `type`
            (`account/invalid-credentials`), so nothing emits `code` today —
            do not branch on it.
        required_acr:
          type: string
          description: For step-up (403) — the ACR the caller must satisfy [D-27].
        enrollment_url:
          type: string
          format: uri
          description: For MFA-enrollment (403) — where to enroll a second factor [D-28].
      additionalProperties: true

    # Legacy error body. Superseded by Problem and emitted by nothing — kept
    # only so generated clients naming the type still compile (see
    # info.description). Referenced by no operation; drop when clients allow.
...
    AssetKind:
      type: string
      enum: [video, audio, image]
    AssetStatus:
      type: string
      enum: [uploaded, processing, ready, failed]
    AssetVisibility:
      type: string
      enum: [private, public]
      default: private
      description: |
        Who may read the asset and its renditions (migration `0032_media_asset_acl`).

        `private` — the owner, plus admins of the owning tenant. This is the
        default and the only value the upload path writes.
        `public` — anyone, including callers with no session; the variant and
        HLS routes serve these anonymously and mark them cacheable by shared
        caches.

        Enforcement is row-level security in Postgres, not handler code: an
        asset the caller may not read does not exist for their query, so every
        such request answers 404 rather than 403.
    Asset:
      type: object
      required: [id, status, kind, mimeType, sizeBytes, createdAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AssetStatus"
        kind:
          $ref: "#/components/schemas/AssetKind"
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        durationMs:
          type: integer
          nullable: true
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        visibility:
          $ref: "#/components/schemas/AssetVisibility"
        hlsUrl:
          type: string
          format: uri
          nullable: true
          description: Master playlist URL; present once `status` is `ready`.
        error:
          type: string
          nullable: true
          description: Failure reason; present when `status` is `failed`.
        createdAt:
          type: string
          format: date-time
...